+++ libs/utils/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [@monodog/utils@1.0.28](https://github.com/manojkmfsi/monodog/releases/tag/@monodog/utils@1.0.28) (2026-03-06)
+
+# @monodog/utils
+
+## 1.0.28
+
+### Patch Changes
+
+
 ## [@monodog/utils@1.0.27](https://github.com/manojkmfsi/monodog/releases/tag/@monodog/utils@1.0.27) (2026-03-06)
 
 # @monodog/utils